UART_WriteByte() add 9-bit support

+++ SWM2X1_StdPeriph_Driver/CSL/SWM2X1_StdPeriph_Driver/SWM2X1_uart.c
@@ -121,11 +121,11 @@ void UART_Close(UART_TypeDef * UARTx)
 * 函数名称:	UART_WriteByte()
 * 功能说明:	发送一个字节数据
 * 输    入: UART_TypeDef * UARTx	指定要被设置的UART串口，可取值包括UART0、UART1、UART4
-*			uint8_t data			要发送的字节			
+*			uint16_t data			要发送的字节
 * 输    出: 无
 * 注意事项: 无
 ******************************************************************************************************************************************/
-void UART_WriteByte(UART_TypeDef * UARTx, uint8_t data)
+void UART_WriteByte(UART_TypeDef * UARTx, uint16_t data)
 {
 	UARTx->DATA = data;
 }
